% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd3pie.R
\name{rd3pie}
\alias{rd3pie}
\title{<Add Title>}
\usage{
rd3pie(Data, Title = NULL, SubTitle = NULL, OuterRadius = NULL,
  width = NULL, height = NULL, enabled_tooltips = NULL, type = NULL)
}
\description{
<Add Description>
}

