% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd3pie.R
\name{rd3pieOutput}
\alias{rd3pieOutput}
\title{Widget output function for use in Shiny}
\usage{
rd3pieOutput(outputId, width = "100\%", height = "100\%")
}
\description{
Widget output function for use in Shiny
}

