% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd3pie.R
\name{renderRd3pie}
\alias{renderRd3pie}
\title{Widget render function for use in Shiny}
\usage{
renderRd3pie(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

